requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
blinker==1.7.0
boto3==1.34.86
botocore==1.34.86
cachetools==5.3.3
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
faiss-cpu==1.8.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
itsdangerous==2.2.0
jmespath==1.0.1
jsonpatch==1.33
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.44
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.48
lxml==5.2.1
marshmallow==3.21.1
multidict==6.0.5
numpy==1.26.4
openai==1.23.2
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pinecone-client==3.2.2
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pypdf==4.2.0
pytz==2024.1
s3transfer==0.10.1
smmap==5.0.1
SQLAlchemy==2.0.29
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tqdm==4.66.2
typing-inspect==0.9.0
tzdata==2024.1
watchdog==4.0.0
Werkzeug==3.0.2
yarl==1.9.4
